→ Moves: Pages card build logic to PagesCardViewModelSlice
