+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/MySiteViewModel.kt
@@ -92,7 +92,7 @@ import org.wordpress.android.ui.mysite.cards.dashboard.CardsTracker
 import org.wordpress.android.ui.mysite.cards.dashboard.bloggingprompts.BloggingPromptsCardAnalyticsTracker
 import org.wordpress.android.ui.mysite.cards.dashboard.domain.DashboardCardDomainUtils
 import org.wordpress.android.ui.mysite.cards.dashboard.domaintransfer.DomainTransferCardViewModel
-import org.wordpress.android.ui.mysite.cards.dashboard.pages.PagesCardContentType
+import org.wordpress.android.ui.mysite.cards.dashboard.pages.PagesCardViewModelSlice
 import org.wordpress.android.ui.mysite.cards.dashboard.plans.PlansCardUtils
 import org.wordpress.android.ui.mysite.cards.dashboard.posts.PostCardType
 import org.wordpress.android.ui.mysite.cards.dashboard.todaysstats.TodaysStatsCardBuilder.Companion.URL_GET_MORE_VIEWS_AND_TRAFFIC
@@ -210,6 +210,7 @@ class MySiteViewModel @Inject constructor(
     private val jetpackFeatureRemovalPhaseHelper: JetpackFeatureRemovalPhaseHelper,
     private val wpJetpackIndividualPluginHelper: WPJetpackIndividualPluginHelper,
     private val blazeCardViewModelSlice: BlazeCardViewModelSlice,
+    private val pagesCardViewModelSlice: PagesCardViewModelSlice,
     private val domainTransferCardViewModel: DomainTransferCardViewModel
 ) : ScopedViewModel(mainDispatcher) {
     private var isDefaultTabSet: Boolean = false
@@ -296,6 +297,7 @@ class MySiteViewModel @Inject constructor(
         _onNavigation,
         siteStoriesHandler.onNavigation,
         blazeCardViewModelSlice.onNavigation,
+        pagesCardViewModelSlice.onNavigation,
         domainTransferCardViewModel.onNavigation
     )
     val onMediaUpload = _onMediaUpload as LiveData<Event<MediaModel>>
@@ -611,10 +613,8 @@ class MySiteViewModel @Inject constructor(
                     onHideMenuItemClick = this::onDashboardCardPlansHideMenuItemClick,
                     onMoreMenuClick = this::onDashboardCardPlansMoreMenuClick
                 ),
-                pagesCardBuilderParams = PagesCardBuilderParams(
-                    pageCard = cardsUpdate?.cards?.firstOrNull { it is PagesCardModel } as? PagesCardModel,
-                    onPagesItemClick = this::onPagesItemClick,
-                    onFooterLinkClick = this::onPagesCardFooterLinkClick
+                pagesCardBuilderParams = pagesCardViewModelSlice.getPagesCardBuilderParams(
+                    cardsUpdate?.cards?.firstOrNull { it is PagesCardModel } as? PagesCardModel,
                 ),
                 activityCardBuilderParams = ActivityCardBuilderParams(
                     activityCardModel = cardsUpdate?.cards?.firstOrNull {
@@ -687,46 +687,6 @@ class MySiteViewModel @Inject constructor(
                 jetpackSwitchMenu = jetpackSwitchMenu
             )
         )
-    }
-
-    private fun onPagesItemClick(params: PagesCardBuilderParams.PagesItemClickParams) {
-        cardsTracker.trackPagesItemClicked(params.pagesCardType)
-        _onNavigation.value = Event(getNavigationActionForPagesItem(params.pagesCardType, params.pageId))
-    }
-
-    private fun getNavigationActionForPagesItem(
-        pagesCardType: PagesCardContentType,
-        pageId: Int
-    ): SiteNavigationAction {
-        return when (pagesCardType) {
-            PagesCardContentType.SCHEDULED -> {
-                SiteNavigationAction.OpenPagesScheduledTab(
-                    requireNotNull(selectedSiteRepository.getSelectedSite()),
-                    pageId
-                )
-            }
-
-            PagesCardContentType.DRAFT -> {
-                SiteNavigationAction.OpenPagesDraftsTab(
-                    requireNotNull(selectedSiteRepository.getSelectedSite()),
-                    pageId
-                )
-            }
-
-            PagesCardContentType.PUBLISH -> {
-                SiteNavigationAction.OpenPages(requireNotNull(selectedSiteRepository.getSelectedSite()))
-            }
-        }
-    }
-
-    private fun onPagesCardFooterLinkClick() {
-        cardsTracker.trackPagesCardFooterClicked()
-        _onNavigation.value =
-            Event(
-                SiteNavigationAction.TriggerCreatePageFlow(
-                    requireNotNull(selectedSiteRepository.getSelectedSite())
-                )
-            )
     }
 
     private fun onActivityCardItemClick(activityCardItemClickParams: ActivityCardItemClickParams) {
